Revert "fix UI"
This reverts commit d2104e6597777d284509aed3ce8cff1c85d286fe.

+++ GentlemanChoice_AndroidStudio/app/src/main/java/com/example/GentlemenChoice/activity/customer/OrderCustomerDetailActivity.java
@@ -172,6 +172,6 @@ public class OrderCustomerDetailActivity extends AppCompatActivity {
         for (CartItem item : items) {
             totalPrice += item.getProductVIew().getPrice() * item.getQuantity();
         }
-        totalPriceTextView.setText(String.format("%d VND", (int) totalPrice));
+        totalPriceTextView.setText(String.format("%.2f VND", totalPrice));
     }
 }
